+++ src/main/java/ThMod_FnH/cards/Marisa/GalacticHalo.java
@@ -24,7 +24,7 @@ public class GalacticHalo extends CustomCard {
   private static final int COST = 2;
   private static final int STC = 2;
   private static final int UPG_STC = 1;
-  private static final int BLC = 13;
+  private static final int BLC = 12;
   private static final int UPG_BLC = 2;
 
   public GalacticHalo() {
