+++ src/main/java/ThMod_FnH/cards/Marisa/StarlightTyphoon.java
@@ -10,7 +10,7 @@ import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import basemod.abstracts.CustomCard;
 import ThMod_FnH.ThMod;
-import ThMod_FnH.cards.special.Spark;
+import ThMod_FnH.cards.derivations.Spark;
 import ThMod_FnH.patches.AbstractCardEnum;
 
 public class StarlightTyphoon extends CustomCard {
