+++ src/main/java/ThMod_FnH/cards/Marisa/WitchOfGreed.java
@@ -21,7 +21,7 @@ public class WitchOfGreed extends CustomCard {
   public static final String NAME = cardStrings.NAME;
   public static final String DESCRIPTION = cardStrings.DESCRIPTION;
   private static final int COST = 1;
-  private static final int STC = 20;
+  private static final int STC = 15;
   private static final int UPG_STC = 10;
 
   private static final int AMP = 1;
