+++ src/main/java/ThMod_FnH/cards/derivations/Spark.java
@@ -1,70 +1,68 @@
-package ThMod_FnH.cards.Marisa;
+package ThMod_FnH.cards.derivations;
 
 import com.megacrit.cardcrawl.actions.AbstractGameAction;
-import com.megacrit.cardcrawl.actions.common.DamageAllEnemiesAction;
-import com.megacrit.cardcrawl.actions.unique.ExhaustAllNonAttackAction;
+import com.megacrit.cardcrawl.actions.common.DamageAction;
 import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
-
-import ThMod_FnH.patches.AbstractCardEnum;
 import basemod.abstracts.CustomCard;
+import ThMod_FnH.action.SparkCostAction;
+import ThMod_FnH.patches.AbstractCardEnum;
 
-public class DarkSpark
-    extends CustomCard {
+public class Spark extends CustomCard {
 
-  public static final String ID = "DarkSpark";
+  public static final String ID = "Spark";
   private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
   public static final String NAME = cardStrings.NAME;
   public static final String DESCRIPTION = cardStrings.DESCRIPTION;
-  public static final String IMG_PATH = "img/cards/Strike.png";
+  public static final String IMG_PATH = "img/cards/Spark.png";
+  private static final int COST = 0;
+  private static final int ATTACK_DMG = 4;
+  private static final int UPGRADE_PLUS_DMG = 2;
 
-  private static final int COST = 1;
-  private static final int ATK_DMG = 14;
-  private static final int UPG_DMG = 4;
-
-  public DarkSpark() {
+  public Spark() {
     super(
         ID,
         NAME,
         IMG_PATH,
         COST,
         DESCRIPTION,
-        CardType.ATTACK,
-        AbstractCardEnum.MARISA_COLOR,
-        CardRarity.UNCOMMON,
-        CardTarget.ALL_ENEMY
+        AbstractCard.CardType.ATTACK,
+        AbstractCardEnum.MARISA_DERIVATIONS,
+        AbstractCard.CardRarity.SPECIAL,
+        AbstractCard.CardTarget.ENEMY
     );
 
-    this.baseDamage = ATK_DMG;
-    this.isMultiDamage = true;
+    this.exhaust = true;
+    this.baseDamage = ATTACK_DMG;
+
   }
 
   public void use(AbstractPlayer p, AbstractMonster m) {
     AbstractDungeon.actionManager.addToBottom(
-        new ExhaustAllNonAttackAction()
+        new DamageAction(
+            m,
+            new DamageInfo(p, this.damage, this.damageTypeForTurn),
+            AbstractGameAction.AttackEffect.SLASH_DIAGONAL
+        )
     );
     AbstractDungeon.actionManager.addToBottom(
-        new DamageAllEnemiesAction(
-            p,
-            this.multiDamage,
-            this.damageTypeForTurn,
-            AbstractGameAction.AttackEffect.NONE
-        )
+        new SparkCostAction()
     );
   }
 
   public AbstractCard makeCopy() {
-    return new DarkSpark();
+    return new Spark();
   }
 
   public void upgrade() {
     if (!this.upgraded) {
       upgradeName();
-      upgradeDamage(UPG_DMG);
+      upgradeDamage(UPGRADE_PLUS_DMG);
     }
   }
 }
