+++ src/main/java/ThMod_FnH/relics/MiniHakkero.java
@@ -4,6 +4,7 @@ import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
 import com.megacrit.cardcrawl.actions.common.RelicAboveCreatureAction;
 import com.megacrit.cardcrawl.actions.utility.UseCardAction;
 import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.helpers.ImageMaster;
 import com.megacrit.cardcrawl.relics.AbstractRelic;
@@ -38,11 +39,33 @@ public class MiniHakkero extends CustomRelic {
 
   public void onUseCard(AbstractCard card, UseCardAction action) {
     flash();
-    ThMod.logger.info("Applying ChargeUpPower");
-    AbstractDungeon.actionManager.addToTop(
-        new ApplyPowerAction(AbstractDungeon.player, AbstractDungeon.player,
-            new ChargeUpPower(AbstractDungeon.player, 1), 1));
-    AbstractDungeon.actionManager
-        .addToBottom(new RelicAboveCreatureAction(AbstractDungeon.player, this));
+    AbstractPlayer p = AbstractDungeon.player;
+    Boolean available = true;
+    int div = 8;
+    if (p.hasRelic("SimpleLauncher")) {
+      div = 6;
+    }
+    if (p.hasPower("ChargeUpPower")) {
+      if (p.getPower("ChargeUpPower").amount >= div) {
+        available = false;
+      }
+    }
+    if (available) {
+      ThMod.logger.info("MiniHakkero : Applying ChargeUpPower for using card : " + card.cardID);
+      AbstractDungeon.actionManager.addToTop(
+          new ApplyPowerAction(
+              AbstractDungeon.player,
+              AbstractDungeon.player,
+              new ChargeUpPower(AbstractDungeon.player, 1),
+              1
+          )
+      );
+    }
+
+    AbstractDungeon.actionManager.addToBottom(
+        new RelicAboveCreatureAction(
+            AbstractDungeon.player, this
+        )
+    );
   }
 }
