+++ src/main/resources/ModTheSpire.json
@@ -11,6 +11,6 @@
   "dependencies": [
     "basemod"
   ],
-  "version": "0.12.18",
+  "version": "0.13.0",
   "sts_version": "09-27-2018"
 }
