+++ src/main/resources/localization/ThMod_Fnh_relics-zh.json
@@ -3,7 +3,7 @@
     "NAME": "迷你八卦炉",
     "FLAVOR": "霖之助的发明。在魔理沙离开人里去往魔法森林时，被霖之助作为礼物送给了魔理沙。",
     "DESCRIPTIONS": [
-      "每打出一张卡，获得一层 #y蓄力 。"
+      "每打出一张卡，获得一层 #y蓄力 。 NL 蓄力超过 #b8 层时失效。"
     ]
   },
   "EnhancedHakkero": {
