+++ src/main/resources/localization/ThMod_Fnh_relics.json
@@ -3,7 +3,7 @@
     "NAME": "Mini-Hakkero",
     "FLAVOR": "Aka mini-EightTrigram Furnace,is originally gift by Rinosuke when Marisa left her parents.",
     "DESCRIPTIONS": [
-      "When you play a card,gain 1  #yCharge-up ."
+      "When you play a card,gain #b1  #yCharge-up .Lose its effect when your #yCharge-Up stacks is higher than #b8 ."
     ]
   },
   "EnhancedHakkero": {
